Move `ToolsSource` to `init`
