+++ lib/codeql.test.js
@@ -42,7 +42,7 @@ const api = __importStar(require("./api-client"));
 const codeql = __importStar(require("./codeql"));
 const defaults = __importStar(require("./defaults.json"));
 const feature_flags_1 = require("./feature-flags");
-const init_action_1 = require("./init-action");
+const init_1 = require("./init");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
 const testing_utils_1 = require("./testing-utils");
@@ -156,7 +156,7 @@ function mockApiDetails(apiDetails) {
             const result = await codeql.setupCodeQL(url, sampleApiDetails, tmpDir, util.GitHubVariant.DOTCOM, false, SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
             t.assert(toolcache.find("CodeQL", `0.0.0-${version}`));
             t.is(result.toolsVersion, `0.0.0-${version}`);
-            t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+            t.is(result.toolsSource, init_1.ToolsSource.Download);
             t.not(result.toolsDownloadDurationMs, undefined);
         }
         t.is(toolcache.findAllVersions("CodeQL").length, 2);
@@ -176,7 +176,7 @@ function mockApiDetails(apiDetails) {
         const result = await codeql.setupCodeQL(url, sampleApiDetails, tmpDir, util.GitHubVariant.DOTCOM, false, SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
         t.assert(toolcache.find("CodeQL", "0.0.0-20200610"));
         t.deepEqual(result.toolsVersion, "0.0.0-20200610");
-        t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+        t.is(result.toolsSource, init_1.ToolsSource.Download);
         t.not(result.toolsDownloadDurationMs, undefined);
     });
 });
@@ -211,7 +211,7 @@ for (const { cliVersion, expectedToolcacheVersion, } of EXPLICITLY_REQUESTED_BUN
             t.assert(releaseApiMock.isDone(), "Releases API should have been called");
             t.assert(toolcache.find("CodeQL", expectedToolcacheVersion));
             t.deepEqual(result.toolsVersion, cliVersion);
-            t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+            t.is(result.toolsSource, init_1.ToolsSource.Download);
             t.not(result.toolsDownloadDurationMs, undefined);
         });
     });
@@ -269,11 +269,11 @@ for (const { isCached, tagName, toolcacheCliVersion } of [
             const result = await codeql.setupCodeQL(undefined, sampleApiDetails, tmpDir, util.GitHubVariant.DOTCOM, false, SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
             t.is(result.toolsVersion, SAMPLE_DEFAULT_CLI_VERSION.cliVersion);
             if (isCached) {
-                t.is(result.toolsSource, init_action_1.ToolsSource.Toolcache);
+                t.is(result.toolsSource, init_1.ToolsSource.Toolcache);
                 t.is(result.toolsDownloadDurationMs, undefined);
             }
             else {
-                t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+                t.is(result.toolsSource, init_1.ToolsSource.Download);
                 t.not(result.toolsDownloadDurationMs, undefined);
             }
         });
@@ -294,7 +294,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
                 variant,
             }, (0, logging_1.getRunnerLogger)(true), false);
             t.deepEqual(result.toolsVersion, "0.0.0-20200601");
-            t.is(result.toolsSource, init_action_1.ToolsSource.Toolcache);
+            t.is(result.toolsSource, init_1.ToolsSource.Toolcache);
             t.is(result.toolsDownloadDurationMs, undefined);
             const cachedVersions = toolcache.findAllVersions("CodeQL");
             t.is(cachedVersions.length, 1);
@@ -317,7 +317,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
                 variant,
             }, (0, logging_1.getRunnerLogger)(true), false);
             t.deepEqual(result.toolsVersion, defaults.cliVersion);
-            t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+            t.is(result.toolsSource, init_1.ToolsSource.Download);
             t.not(result.toolsDownloadDurationMs, undefined);
             const cachedVersions = toolcache.findAllVersions("CodeQL");
             t.is(cachedVersions.length, 2);
@@ -337,7 +337,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
         });
         const result = await codeql.setupCodeQL("latest", sampleApiDetails, tmpDir, util.GitHubVariant.DOTCOM, false, SAMPLE_DEFAULT_CLI_VERSION, (0, logging_1.getRunnerLogger)(true), false);
         t.deepEqual(result.toolsVersion, defaults.cliVersion);
-        t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+        t.is(result.toolsSource, init_1.ToolsSource.Download);
         t.not(result.toolsDownloadDurationMs, undefined);
         const cachedVersions = toolcache.findAllVersions("CodeQL");
         t.is(cachedVersions.length, 2);
@@ -374,7 +374,7 @@ for (const variant of [util.GitHubVariant.GHAE, util.GitHubVariant.GHES]) {
             tagName: defaults.bundleVersion,
             variant: util.GitHubVariant.GHAE,
         }, (0, logging_1.getRunnerLogger)(true), false);
-        t.is(result.toolsSource, init_action_1.ToolsSource.Download);
+        t.is(result.toolsSource, init_1.ToolsSource.Download);
         t.not(result.toolsDownloadDurationMs, undefined);
         const cachedVersions = toolcache.findAllVersions("CodeQL");
         t.is(cachedVersions.length, 1);
