+++ lib/init-action.js
@@ -23,7 +23,6 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.ToolsSource = void 0;
 const path = __importStar(require("path"));
 const core = __importStar(require("@actions/core"));
 const actions_util_1 = require("./actions-util");
@@ -37,13 +36,6 @@ const repository_1 = require("./repository");
 const trap_caching_1 = require("./trap-caching");
 const util_1 = require("./util");
 const workflow_1 = require("./workflow");
-var ToolsSource;
-(function (ToolsSource) {
-    ToolsSource["Unknown"] = "UNKNOWN";
-    ToolsSource["Local"] = "LOCAL";
-    ToolsSource["Toolcache"] = "TOOLCACHE";
-    ToolsSource["Download"] = "DOWNLOAD";
-})(ToolsSource = exports.ToolsSource || (exports.ToolsSource = {}));
 async function sendInitStatusReport(actionStatus, startedAt, config, toolsDownloadDurationMs, toolsFeatureFlagsValid, toolsSource, toolsVersion, logger) {
     const statusReportBase = await (0, actions_util_1.createStatusReportBase)("init", actionStatus, startedAt);
     const languages = config.languages.join(",");
@@ -80,9 +72,9 @@ async function sendInitStatusReport(actionStatus, startedAt, config, toolsDownlo
         trap_cache_languages: Object.keys(config.trapCaches).join(",") || "",
         trap_cache_download_size_bytes: Math.round(await (0, trap_caching_1.getTotalCacheSize)(config.trapCaches, logger)) || -1,
         trap_cache_download_duration_ms: Math.round(config.trapCacheDownloadTime) || -1,
-        tools_source: toolsSource || ToolsSource.Unknown,
+        tools_source: toolsSource || init_1.ToolsSource.Unknown,
     };
-    if (toolsSource !== ToolsSource.Download) {
+    if (toolsSource !== init_1.ToolsSource.Download) {
         await (0, actions_util_1.sendStatusReport)(statusReport);
     }
     // Otherwise, we should append the extra two download-related telemetry fields.
