+++ lib/setup-codeql.js
@@ -40,7 +40,7 @@ const api = __importStar(require("./api-client"));
 // creation scripts. Ensure that any changes to the format of this file are compatible with both of
 // these dependents.
 const defaults = __importStar(require("./defaults.json"));
-const init_action_1 = require("./init-action");
+const init_1 = require("./init");
 const util = __importStar(require("./util"));
 const util_1 = require("./util");
 exports.CODEQL_DEFAULT_ACTION_REPOSITORY = "github/codeql-action";
@@ -478,19 +478,19 @@ async function setupCodeQLBundle(toolsInput, apiDetails, tempDir, variant, bypas
     switch (source.sourceType) {
         case "local":
             codeqlFolder = await toolcache.extractTar(source.codeqlTarPath);
-            toolsSource = init_action_1.ToolsSource.Local;
+            toolsSource = init_1.ToolsSource.Local;
             break;
         case "toolcache":
             codeqlFolder = source.codeqlFolder;
             logger.debug(`CodeQL found in cache ${codeqlFolder}`);
-            toolsSource = init_action_1.ToolsSource.Toolcache;
+            toolsSource = init_1.ToolsSource.Toolcache;
             break;
         case "download": {
             const result = await downloadCodeQL(source.codeqlURL, source.cliVersion, apiDetails, variant, tempDir, logger);
             toolsVersion = result.toolsVersion;
             codeqlFolder = result.codeqlFolder;
             toolsDownloadDurationMs = result.toolsDownloadDurationMs;
-            toolsSource = init_action_1.ToolsSource.Download;
+            toolsSource = init_1.ToolsSource.Download;
             break;
         }
         default:
